+++ src/check.ts
@@ -19,7 +19,7 @@ import {nakedDomain} from "./util/stringUtils";
 
 const supportedDataFields = new Map<Target, DataField[]>([
     [Target.Scene, [DataField.Id, DataField.Title, DataField.Organized, DataField.Studio, DataField.Code, DataField.Date, DataField.Tags, DataField.Files]],
-    [Target.Performer, [DataField.Id, DataField.Name, DataField.Disambiguation, DataField.Favorite, DataField.AliasList, DataField.Birthdate, DataField.HeightCm, DataField.Tags]],
+    [Target.Performer, [DataField.Id, DataField.Name, DataField.Disambiguation, DataField.Favorite, DataField.AliasList, DataField.Birthdate, DataField.CareerLength, DataField.HeightCm, DataField.Tags]],
     [Target.Gallery, [DataField.Id, DataField.Title, DataField.Date, DataField.Tags, DataField.Files]],
     [Target.Group, [DataField.Id, DataField.Name, DataField.Aliases, DataField.Date, DataField.Studio, DataField.Director, DataField.SceneCount]],
     [Target.Studio, [DataField.Id, DataField.Name, DataField.Aliases]],